requirements.txt
email
spacy
sentence-transformers
pinecone-client
flask
